requirements.txt
streamlit
openai
requests
ffmpeg-python
streamlit-audio-recorder
streamlit run app.py
